feat(model): implement class lotto
Added validate methods:
- is valid length
- is all unique
- is all in range
